content/media/video/test/test_ended1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: ended</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogg'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test if the ended event works correctly.
var v = document.getElementById('v');
var endPassed = false;
var completed = false;
var timeout;

function startTest() {
  if (completed)
    return false;

  v.play();
  return false;
}

function playbackEnded() {
  if (completed)
    return false

  completed = true;
  //  The following check is disabled and will be re-enabled by bug 460871
  // ok(v.currentTime >= 3.9 && v.currentTime <= 4.0,
  //    "Checking currentTime at end: " + v.currentTime);
  ok(v.ended, "Checking playback has ended");
  clearTimeout(timeout);
  SimpleTest.finish();
  return false;
}
timeout = setTimeout(function () {
                       ok(false, "Test timed out");
                       SimpleTest.finish();
                     }, 30000);

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
